Extract FailureOnlyReporter to new file and use as base of reporter for cross-browser tests
